Refactor, refactor code structure, fix pynq rpc (#29)
